requirements.txt
streamlit
pandas
plotly
pycryptodome
bcrypt
scikit-learn